+++ projects/ontimize-web-ngx/src/lib/interfaces/o-table-onclick.interface.ts
@@ -1,0 +1,7 @@
+export interface OnClickTableEvent {
+  /** row data */
+  row: any;
+  /** row index */
+  rowIndex: number;
+  mouseEvent: MouseEvent;
+}